% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addNoMS2.R
\name{addNoMS2}
\alias{addNoMS2}
\title{add no MS2 data to compMS2 class object internal to corrNetwork function}
\usage{
addNoMS2(object, ...)
}
\arguments{
\item{object}{a "compMS2" class object.}

\item{specNames}{character vector of composite spectrum names.}

\item{eicMzRt}{matrix of EICnos/unique id, mz values and rt values in 3 columns.}
}
\value{
a "compMS2" class object with noMS2 data added to the appropriate slots.
}
\description{
add no MS2 data to compMS2 class object internal to corrNetwork function
}
