% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adduct2mass.R
\name{adduct2mass}
\alias{adduct2mass}
\title{converts ESI adduct names into table of monoisotopic masses}
\source{
element monoisotopic masses and natural abundances taken from \url{http://www.sisweb.com/referenc/source/exactmas.htm}, see ?exactMassEle.
}
\usage{
adduct2mass(adductNames = c("[M-H]-", "[2M+2CH3OH]2-", "[M-H+C2H4O2+Na]-"))
}
\arguments{
\item{adductNames}{character vector of ESI adduct names in a specific form for example see default names c('[M-H]-', '[2M+2CH3OH]2-', '[M-H+C2H4O2+Na]-'). The 
function was developed and tested against the 128 and 133 different ESI 
adducts and in-source fragment names for accuracy.}
}
\value{
data.frame of n rows corresponding to each adduct name and 5 columns:
\enumerate{
\item "name" adduct name    
\item "nmol" number of molecules (e.g. 2M-H = 2, M-H = 1)   
\item "Ch" charge state      
\item "massDiff" summed monoisotopic mass difference.
\item "mode" ion polarity based on adduct name (e.g. "]-" = 'neg", and "]+" = 'pos")
}
}
\description{
converts ESI adduct names into table of monoisotopic masses
}
\details{
this function can be used to generate a table of ESI adducts for the
\code{\link{metID.dbAnnotate}} function.
}
\references{
\enumerate{
\item Stanstrup, J., Gerlich, M., Dragsted, L.O. et al. 
Anal Bioanal Chem (2013) 405: 5037. doi:10.1007/s00216-013-6954-6
}
}
