% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annoCompoundClass.R
\name{annoCompoundClass}
\alias{annoCompoundClass}
\title{automatically annotate compound classes comments table}
\usage{
annoCompoundClass(object, ...)
}
\arguments{
\item{object}{a compMS2 class object.}

\item{overWrite}{logical should any existing compound_class types be overwritten in the
comments table (default = FALSE).}

\item{minSimScore}{numeric mimimum chemical similarity score (default = 0.8),
all chemical fingerprints with similarity above this score to entries of HMDB,
drugBank and T3DB will be considered for the compound class. The highest 
chemical similarity score above this threshold will be considered as the 
predicted compounds class. In the event of multiple tied maximum chemical
similarity scores the most frequent compound class will added to the comments
table.}
}
\value{
the compMS2 class object with compound classes (currently only phospholipids supported) contained in
data(lipidAbbrev) annotated and interpreted lipid tail information added to the Comments table
accessible by Comments(object).

a compMS2 class object with the compound class of the identified
compound added to the metID comments table. Comments(object). Additionally
a barplot is generated summarizing the identified compound class information
returned.
}
\description{
automatically annotate compound classes comments table
}
\details{
This function attempts to automatically add chemical taxonomy information
to the metID comments table accessible using Comments(object). Pre-existing
compound class information can be overwritten. Additionally a predicted
compound class is added for any compound not contained in HMDB, drugBank or
T3DB above a minimum chemical similarity score.
}
