% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanAbs.R
\name{cleanAbs}
\alias{cleanAbs}
\title{Adapted from PubMedWordCloud (cleanAbstracts) to work with compMS2Miner}
\usage{
cleanAbs(Abs, rmNum = TRUE, tolw = TRUE, toup = FALSE, rmWords = TRUE,
  yrWords = NULL, stemDoc = FALSE)
}
\arguments{
\item{Abs}{output of getAbs, or just a paragraph of text}
}
\description{
Adapted from PubMedWordCloud (cleanAbstracts) to work with compMS2Miner
}
