% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combineMS2.Ions.R
\name{combineMS2.Ions}
\alias{combineMS2.Ions}
\title{Combine spectra peaks within individual spectra}
\usage{
combineMS2.Ions(object, ...)
}
\arguments{
\item{...}{option arguments to be passed along.}

\item{mzError}{interpeak absolute m/z error for spectra signal grouping (default = 0.01).}

\item{minPeaks}{Minimum number of peaks per spectrum (default = 3).}
}
\value{
A compMS2 object with ion grouped composite spectra.
}
\description{
Combine spectra peaks within individual spectra
}
\details{
group ions according to absolute m/z error. 
The default parameters are suitable for a high-resolution Q-ToF.
Following ion grouping, signal intensities are summed and an average m/z 
calculated for each ion group. This signal summing serves to increase the
overall intensity of true ion signal across multiple scans and reduce the 
contribution of noise within the spectrum. Calculation of the central tendency
of each ion group serve to homogenize the random error and improve the mass
accuracy of each spectrum peak.
}

