% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combineMS2.removeContam.R
\name{combineMS2.removeContam}
\alias{combineMS2.removeContam}
\title{remove any possible contaminants identified as repeating sequences of isobaric ions of 
high spectral similarity.}
\usage{
combineMS2.removeContam(object, ...)
}
\arguments{
\item{object}{a "compMS2" class object.}

\item{ms1Abs}{numeric ms1 mass to charge absolute error for hierarchical 
clustering of MS1 masses (default = 0.01). Utilizes the median method of the
\code{\link{hclust}} function of the fastcluster package.}

\item{maxRtGap}{numeric maximum retention time gap (in seconds) between two 
isobaric ions. Sequences of possible contaminants will be identified using this
difference in retention time (default = 60).}

\item{nContams}{numeric number of isobaric ions in a sequence separated by a maximum
retention time gap (maxRtGap). Any sequences of isobaric ions greater than or equal 
to this number will be removed (default = 10).}

\item{minSimScore}{numeric minimum spectral similarity score (values between 0-1).
If any isobar in a sequence of possible contaminants is below this minimum 
mean dot product similarity score
then it will not be removed. This is to ensure that only true isobaric contaminants
are removed and spectra which have been grouped amongst them are not erroneously
removed (default = 0.8).}

\item{remContam}{logical should possible contaminant spectra be automatically
removed from the object (default = TRUE), If FALSE the contaminant plot
will still be printed but the spectra will not be removed. In this way the user
can interactively determine suitable parameters prior to spectrum removal.}
}
\value{
a "compMS2" class object with the composite spectra and all metID 
information of any contaminants identified removed. Any correlation or spectral
similarity networks will also have to be recalculated.
}
\description{
remove any possible contaminants identified as repeating sequences of isobaric ions of 
high spectral similarity.
}
\details{
The function produces plots to visualize the contaminants identified.
If you suspect all isobaric ions across the gradient are contaminants then
you can set the argument maxRtGap to Infinite (maxRtGap=Inf). N.B. Any spectrum which is not sufficiently similar (specSimScore) to the rest
of the potential contaminants across the gradient will not be erroneously removed.
}
\examples{
compMS2contamRem <- combineMS2(compMS2Example, 'removeContam', maxRtGap=Inf, 
                               nContams=4)
}
