% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compMS2.R
\name{compMS2}
\alias{compMS2}
\title{Constructor for compMS2 class object from a peak table and MS2 mzXML file(s)}
\usage{
compMS2(MS1features = NULL, mzXMLdir = NULL, nSlaves = NULL,
  mode = "pos", precursorPpm = 10, ret = 10, TICfilter = 10000)
}
\arguments{
\item{nSlaves}{Number of cores for parallel computation?}

\item{mode}{Ionisation polarity ?}

\item{precursorPpm}{Parts per million mass accuracy to match MS1 features to MS2 spectra (ppm)}

\item{ret}{Retention time tolerance to match MS1 features to MS2 spectra (+/- seconds)}

\item{TICfilter}{Minimum Total Ion Current to consider an MS2 spectrum ? 
#@param pH Approximate pH of mobile phase for LogD of DB matches 
#@param minTanimotoScore minimum tanimoto chemical similarity score?}
}
\value{
A compMS2 object
}
\description{
Matches MS1 features to MS2 spectra (.mzXML) files based on a 
mass-to-charge and retention time tolerance. Composite spectra and other data
can subsequently be visualized during any stage of the CompMS2miner
processing workflow. Composite spectra can be denoised, ion signals grouped 
and summed, substructure groups identified, common Phase II metabolites
predicted and features matched to data bases monoisotopic mass data 
and insilico MS2 fragmentation data.
The resulting data can then be readily curated by sending to a local or online
couchDB database.
}

