% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compMS2Construct.R
\name{compMS2Construct}
\alias{compMS2Construct}
\title{Constructor for compMS2 class object from a peak table and MS2 mzXML/mzML/.mgf
file(s)}
\usage{
compMS2Construct(MS1features = NULL, msDataDir = NULL, MS2files = NULL,
  nCores = NULL, argsCorrNetwork = list(obsNames = NULL, corrThresh = 0.6,
  corrMethod = "spearman", delta = 0.05, MTC = "BH"), mode = "pos",
  precursorPpm = 10, ret = 10, TICfilter = 10000, minPeaks = 1,
  isoWid = 4, verbose = TRUE)
}
\arguments{
\item{MS1features}{either a data.frame, full file path as a character string to a  .csv file of a MS1 feature table in the form observation (samples) in columns and
variables (Mass spectral signals) in rows, the first 3 columns must consist of:
\enumerate{
 \item EIC number or unique peak identifier.
 \item mass-to-charge ratio of peak group.
 \item median/ peak apex retention time in seconds. 
 }
 
 optionally the 4th column of the MS1feature table may contain any adducts
 and isotopes identified by for example the CAMERA R package.
 If this column is present the adducts will be incorporated in to the compMS2
 class object and used to guide the subsequent \code{\link{metID.dbAnnotate}}
 function. This can be very useful for narrowing possible annotations in
 subsequent stages of the compMS2miner workflow and particularly in reduction
 of false positives annotations. The adduct annotations must
 consist of the following notation style for example [M-H]-, [2M+2CH3OH]2-, 
 [M-H+C2H4O2+Na]-. Abbreviations such as Hac (CH3COOH) for acetic acid 
 and ACN (i.e. C2H3N) for acetonitrile 
 are not acceptable formulae must be used to determine the correct
 elemental composition must be included. As is typical of the output of
 CAMERA for example multiple possible adducts can appear for the same feature
 where they have shared/similar expected masses.
   
 If argument is not supplied a GUI (tcltk) file selection window will open and a .csv file can then be selected.}

\item{msDataDir}{character full path to a directory containing LC-MS/MS data files
in either the open framework .mzXML or newer .mzML file types also mascot generic format files (.mgf). If argument is
not supplied a GUI (tcltk) file selection window will open and the directory 
can be selected.}

\item{MS2files}{character vector of full paths to ms2 files (either .mzML, .mzXML or .mgf).
Alternative to choosing the directory. In this way particular files within a directory
or files from multiple directory locations can be specified.}

\item{nCores}{numeric Number of cores for parallel computation.}

\item{mode}{character Ionisation polarity must be either 'pos' or 'neg'.}

\item{precursorPpm}{numeric Parts per million mass accuracy to match MS1 features to MS2 spectra (ppm)}

\item{ret}{numeric retention time tolerance to match MS1 features to MS2 spectra (+/- seconds).}

\item{TICfilter}{numeric Minimum Total Ion Current to consider an MS2 spectrum. Any MS2 scan
below this threshold will not be considered.}

\item{isoWid}{numeric isolation width of DDA precursor ions, utilized to identify
potentially chimeric spectra.}

\item{verbose}{logical if TRUE display progress bars.}

\item{minPeaks.}{minimum number of fragment ions for a spectrum to be 
considered (default = 1).}
}
\value{
A compMS2 object
}
\description{
Matches MS1 features to MS2 spectra (.mzXML/.mzML/.mgf) files 
based on a mass-to-charge and retention time tolerance. Composite spectra 
and other data can subsequently be visualized during any stage of the compMS2Miner
processing workflow. Composite spectra can be denoised, ion signals grouped 
and summed, substructure groups identified, common Phase II metabolites
predicted and features matched to data bases monoisotopic mass data 
and insilico MS2 fragmentation data.
The resulting data can then be readily curated by sending to a local or online
couchDB database.
}
