% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/compMS2create.R
\name{compMS2create}
\alias{compMS2create}
\title{create a compMS2 object}
\usage{
compMS2create(MS2file = NULL, MS1features = NULL, TICfilter = 10000,
  precursorPpm = 10, ret = 10)
}
\arguments{
\item{MSfiles}{character vector of mzXML file locations}
}
\description{
create a compMS2 object
}

