% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compMS2Explorer.R
\name{compMS2Explorer}
\alias{compMS2Explorer}
\title{Visualize your compMS2Miner results output using a shiny app.}
\usage{
compMS2Explorer(object, ...)
}
\arguments{
\item{object}{a compMS2 class object or a character full path to a compMS2Miner zip file}

\item{browserLaunch}{logical launch app in web browser (default = TRUE).}
}
\description{
Visualize your compMS2Miner results output using a shiny app.
}
