% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/couchDBcurate.R
\name{couchDBcurate}
\alias{couchDBcurate}
\title{CouchDB login
send CompMS2 data set to couchDB}
\usage{
couchDBcurate(object, ...)
}
\arguments{
\item{couchDBname}{New or existing CouchDB database name (must be all lower case,  can contain underscores)}
}
\value{
CouchDB records : All Results from the current stage of the CompMS2miner
 are sent to the already established/ newly created couchDB database.
 The following documents are sent to couchDB :
}
\description{
CouchDB login
send CompMS2 data set to couchDB
}

