% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/couchDBpingReq.R
\name{couchDBpingReq}
\alias{couchDBpingReq}
\title{CouchDBpingReq
send ping to couch db using log in credentials from GetLoginDetails or
named character vector}
\usage{
couchDBpingReq(credentials)
}
\arguments{
\item{credentials}{either output from GetLoginDetails or named character vector "host", "Username", "Password" containing user Login parameters (e.g. c( host = "localhost", Username = "", Password = ""))}
}
\value{
ping request result
}
\description{
CouchDBpingReq
send ping to couch db using log in credentials from GetLoginDetails or
named character vector
}

