% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cytoScapeMap.R
\name{cytoScapeMap}
\alias{cytoScapeMap}
\title{Create cytoscape map files based on correlation (Spearman's rho)}
\usage{
cytoScapeMap(object, ...)
}
\description{
Create cytoscape map files based on correlation (Spearman's rho)
}

