% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/deconvNoise.DNF.R
\name{deconvNoise.DNF}
\alias{deconvNoise.DNF}
\title{Spectral noise filtration using dynamic noise filter}
\source{
Xu H. and Frietas M. "A Dynamic Noise Level Algorithm for Spectral
Screening of Peptide MS/MS Spectra" 2010 BMC Bioinformatics.
}
\usage{
deconvNoise.DNF(object, ...)
}
\arguments{
\item{DNF.}{dynamic noise filter minimum signal to noise threshold
(default = 2), calculated as the ratio between the linear model predicted
intensity value and the actual intensity.}

\item{minPeaks.}{minimum number of signal peaks following dynamic
noise filtration (default = 5).}

\item{maxPeaks.}{maximum number of signal peaks the function will continue
until both the minimum DNF signal to noise ratio is exceeding and the number
of peaks is lower than the maximum (default = 5).}
}
\value{
noise filtered composite MS2 spectra.
}
\description{
uses the dynamic noise filtration algorithm adapted from the method described
in Xu H. and Frietas M. "A Dynamic Noise Level Algorithm for Spectral
Screening of Peptide MS/MS Spectra" 2010 BMC Bioinformatics.
}

