% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deconvNoise.R
\name{deconvNoise}
\alias{deconvNoise}
\title{Filter spectral noise from a CompMS2 class object}
\usage{
deconvNoise(object, ...)
}
\arguments{
\item{...}{option arguments to be passed along.}

\item{object.}{a compMS2 class object obtained from the function CompMSset}

\item{method.}{dynamic noise file "DNF" or fixed maximum intensity "maxInt"}
}
\value{
A compMS2 object with noise filtered composite spectra.
}
\description{
Filter spectral noise from a CompMS2 class object
}
\seealso{
\code{\link{deconvNoise.DNF}}, \code{\link{deconvNoise.maxInt}}
}
