% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dotProdMatrix.R
\name{dotProdMatrix}
\alias{dotProdMatrix}
\title{dot product matrix calculation}
\usage{
dotProdMatrix(allSpectra = NULL, spectraNames = NULL, binSizeMS2 = NULL)
}
\arguments{
\item{allSpectra}{a numeric matrix consisting of two columns 1. mass and 2. intensity}

\item{spectraNames}{character names of individual spectra to compare must equal number of rows of allSpectra}

\item{binSizeMS2}{numeric the MS2 bin size to bin MS2 data prior to dot product calculation (default = 0.1 Da).}
}
\value{
a matrix of equal dimension corresponding to the number of factor levels f
}
\description{
dot product matrix calculation
}

