% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exactMassEle.R
\docType{data}
\name{exactMassEle}
\alias{exactMassEle}
\title{elemental monoisotopic masses and natural abundances}
\format{A data frame with 83 rows and 4 columns}
\source{
\url{http://www.sisweb.com/referenc/source/exactmas.htm}
}
\usage{
data(exactMassEle)
}
\description{
This dataset contains data taken from the table available at (\url{http://www.sisweb.com/referenc/source/exactmas.htm})
The variables are as follows:
}
\details{
\itemize{
 \item eleName full element name (e.g. sodium).
 \item eleSymbol element symbol (e.g. Na).
 \item monoMass character of isotope mass(es) seperated by a space.
 \item natAbund character of natural isotope abundance(s) seperated by a space.
 }
}
\keyword{datasets}
