% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/falsePosIdentify.R
\name{falsePosIdentify}
\alias{falsePosIdentify}
\title{Identify and/or remove false positives from metID comments table based on a prohibited
keyword search}
\usage{
falsePosIdentify(object, ...)
}
\arguments{
\item{object}{a "compMS2" class object.}

\item{prohibKeyWords}{character a regular expression of prohibited keywords.
If any of these are discovered in any of the abstracts returned from the
PubMed Entrez search then they will be removed from the metID comments table.}

\item{n}{integer number of pubmedids to return per metID comments possible_identity.
(default = 50). Larger numbers may improve accuracy but increase computation time.
The maximum allowed value is 500 (limited by entrez system).}

\item{meanFreqPerAbs}{numeric minimum mean frequency of the summed frequencies
of the prohibited keywords. Default = 0.6 that is a mean summed frequency of 
any of the key words in the regular expression of 0.6 for it to be
considered. This limits the accidental removals of a true positive which
has a low word count frequency by chance.}

\item{removeFP}{logical if TRUE remove possible false positive annotations
from the possible_identity column of the metID comments table. If FALSE (default) the
possible false positive annotation names will remain but they will be
flagged in the user_comments column of the metID comments table as 
"possible false positive (falsePosRemoval)".}

\item{maxChar}{numeric maximum number of characters in cleaned abstract words to return.}

\item{verbose}{logical if TRUE display progress bars.}
}
\description{
Identify and/or remove false positives from metID comments table based on a prohibited
keyword search
}
