% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/form2Mat.R
\name{form2Mat}
\alias{form2Mat}
\title{mol formula to element count table}
\usage{
form2Mat(atomForm = NULL)
}
\arguments{
\item{atomForm}{a character vector of atomic formulae.}
}
\value{
list containing two matrices one matrix "nEleTable" with a column for each unique element and a row for each atomic
formula with numbers of each element and another matrix 'eleRatio" with the ratios of all elements in each atomic formula.
}
\description{
mol formula to element count table
}
