% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formulaFilter.R
\name{formulaFilter}
\alias{formulaFilter}
\title{filter potential formulae based on maximum element by mass and ratios}
\usage{
formulaFilter(atomForm = NULL, mass = NULL)
}
\arguments{
\item{atomForm}{character vector of atomic formulae}

\item{mass}{numeric the mass of all formulae considered.}
}
\description{
filter potential formulae based on maximum element by mass and ratios
}
\details{
maximum element numbers and element ratios taken from 5 chemical
databases (n=84891 structures)
}
