% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAbs.R
\name{getAbs}
\alias{getAbs}
\title{Adapted from PubMedWordCloud to work with compMS2Miner}
\usage{
getAbs(PMID)
}
\arguments{
\item{PMID}{character vector of pubMed ids to get abstracts for.}
}
\description{
Adapted from PubMedWordCloud to work with compMS2Miner
}
\details{
if the query sequence is too long than 500 this function will not work
}
