% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/getLoginDetails.R
\name{getLoginDetails}
\alias{getLoginDetails}
\title{Get Login details for CouchDB}
\usage{
getLoginDetails()
}
\arguments{
\item{Host}{couchDB host name, defaults to localhost. Can be online repository.}

\item{Name}{couchDB administrator username (used for http commands, not stored).}

\item{Password}{couchDB administrator password (used for http commands, not stored).}
}
\value{
login details for couchDB
}
\description{
tcltk GUI to get login details for couchDB
}

