% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/getTitles.R
\name{getTitles}
\alias{getTitles}
\title{get PubMed title function adapted from PubChemWordcloud package v 0.3.2}
\usage{
getTitles(pmid)
}
\arguments{
\item{pmid}{pubmed id number}
}
\description{
get PubMed title function adapted from PubChemWordcloud package v 0.3.2
}

