% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HMDB.R
\docType{data}
\name{HMDB}
\alias{HMDB}
\title{Human Metabolome Database}
\format{A data frame with 41478 rows and 7 variables}
\source{
\url{http://www.hmdb.ca/downloads}
}
\usage{
data(HMDB)
}
\description{
This dataset contains fields from the HMDB database 
The variables are as follows:
}
\details{
\itemize{
 \item WebAddress web address for database (when combined with the 
 Unique_DB_ID entry this will provide a direct link to the online database
 entry e.g. Web address : "www.hmdb.ca/metabolites/" +
 Unique_DB_ID : "HMDB00001" = \url{www.hmdb.ca/metabolites/HMDB00001})
 \item Unique_DB_ID. HMDB unique reference number (HMDB00001 -- HMDB61388)
 \item name. HMDB entry name
 \item monoisotopic_weight. Monoisotopic weight of HMDB entry 
 (1.007825 -- 6176.017855)
 \item SMILES. Canonical SMILES code of HMDB entry
 \item molecular_formula. Molecular formula of HMDB entry 
 \item cas_number. CAS registry number of HMDB entry
 }
}
\keyword{datasets}

