% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/intFeatCorr.R
\name{intFeatCorr}
\alias{intFeatCorr}
\title{Interfeature correlation}
\usage{
intFeatCorr(object, ...)
}
\arguments{
\item{sampleIDstring}{Unique text string to identify sample columns for correlation matrix calculation}
}
\description{
Interfeature correlation
}

