% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metFragAdducts.R
\docType{data}
\name{metFragAdducts}
\alias{metFragAdducts}
\title{data.frame of customizable metFrag adduct types and codes}
\format{A data frame with 9 rows and 3 columns}
\source{
\url{http://c-ruttkies.github.io/MetFrag/projects/metfragcl/}
}
\usage{
data(metFragAdducts)
}
\description{
This dataset is the default adduct type table for the \code{\link{metID.metFrag}}
function. A custom table can be created following this format as more adduct
types are added in future versions of the metFrag command line tool.
N.B. adduct names must match those supplied to the \code{\link{adduct2mass}} function internal
to the \code{\link{metID.dbAnnotate}} function.
}
\details{
\enumerate{
 \item adduct adduct name string must match that supplied to adduct2mass.
 \item metFragCode the MetFrag command line tool code for the adduct.
 \item mode polarity. must be either 'pos' or 'neg'. 
 }
}
\keyword{datasets}
