% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metFragCl.R
\name{metFragCl}
\alias{metFragCl}
\title{metFrag command line function for localSDF files}
\source{
\url{http://c-ruttkies.github.io/MetFrag/projects/metfragcl/} developed based on the command line .jar file (MetFrag2.3-CL.jar) downloaded on 2016/07/12
\enumerate{
\item MetFrag relaunched: incorporating strategies beyond in silico fragmentation: C Ruttkies, E L Schymanski, S Wolf, J Hollender, S Neumann Journal of Cheminformatics 2016 8:3
\item In silico fragmentation for computer assisted identification of metabolite mass spectra: S Wolf, S Schmidt, M Müller-Hannemann, S Neumann BMC bioinformatics 11 (1), 148
}
}
\usage{
metFragCl(massSpectrum = NULL, precMass = NULL, compSpecName = NULL,
  dbEntryTable = NULL, metFragJar = NULL, SDFtmp = NULL,
  keepTempFiles = FALSE, mode = "pos", frag_mzabs = 0.05,
  esiList = NULL, maxTreeDepth = 1)
}
\arguments{
\item{massSpectrum}{data.frame composite spectrum consisting of two columns mass and intensity.}

\item{precMass}{numeric the MS1 m/z (precursor mass).}

\item{compSpecName}{character name of composite spectrum for directory and file naming.}

\item{dbEntryTable}{data.frame with the requisite information for the SDFtmp localSDF database consisting of at least 4 columns named 1. 'WebAddress', 2. 'DBid', 3. 'DBname', 4. 'SMILES'.}

\item{metFragJar}{character full path to metFragCL.jar file (extdata in compMS2Miner package).}

\item{SDFtmp}{an "SDFset" class object of SDF file for the localSDF database search of metFragCL.}

\item{keepTempFiles}{logical default = FALSE, sdf, mf and results files will
be created as temporary files otherwise if TRUE files will be retained in named subdirectories (see details).}

\item{mode}{character ionization polarity (either 'pos' or 'neg').}

\item{frag_mzabs}{numeric delta predicted-observed fragment mass accuracy for matching.}

\item{esiList}{named numeric vector of electrospray type numbers for metFrag params file. e.g. positive mode 
\tabular{llll}{
\cr M+H \tab M+NH4  \tab M+Na   \tab M+K 
\cr 1    \tab 18    \tab 23    \tab 39
}}

\item{maxTreeDepth}{numeric fragments of fragments? (default = 1 i.e. only direct daughter ions of the structure will be considered). Setting the tree depth to higher values may cause the metFragCL to take longer.}
}
\value{
if MetFrag2.3-CL.jar process completed then a data.frame containing any fragments matched to the composite mass spectra are returned.
MetFragPreProcessingCandidateFilter = UnconnectedCompoundFilter%\n',
}
\description{
metFrag command line function for localSDF files
}
