% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metID.CFM.R
\name{metID.CFM}
\alias{metID.CFM}
\title{Wrapper function for Competitive Fragmentation modelling (CFM) \emph{in silico} fragmentation software}
\source{
\enumerate{
\item Allen F, Pon A, Wilson M, Greiner R, and Wishart D. CFM-ID: a web server for annotation, spectrum prediction and metabolite identification from tandem mass spectra. Nucleic Acids Res. June 2014. \url{http://nar.oxfordjournals.org/content/early/2014/06/03/nar.gku436.full}.
\item fraggraph-gen.exe file in extdata downloaded (2016/07/09, cfm-id-2.2_win32.zip) from \url{https://sourceforge.net/p/cfm-id/wiki/Home/}. 
\item lpsolve.dll in extdata downloaded (2016/07/09, lp_solve_5.5.2.3_IDE_Setup.exe) from \url{https://sourceforge.net/projects/lpsolve/}.
}
}
\usage{
metID.CFM(object, ...)
}
\arguments{
\item{object}{a "compMS2" class object.}

\item{featureSubSet}{character vector of composite spectra names (e.g. CC_1, CC_2 etc.) otherwise the default is to perform CFM fragmentation on all composite spectra.}

\item{keepTempFiles}{logical default = FALSE fraggraph-gen .csv output file will
be created as temporary files otherwise if TRUE file will be retained in subdirectories named by composite spectrum name.}

\item{minPropTicEx}{numeric minimum mean total ion current explained (default = 0.9)
the candidate with the highest proportion of the total ion current explained
 above this minimum will be automatically added to the Comments
table. The argument autoPossId must also be set to TRUE.}

\item{autoPossId}{logical if TRUE the function will automatically add the name
of the top annotation based on mean total ion current explained and metFrag score
 (default = FALSE). Caution if TRUE
this will overwrite any existing possible_identities in the "metID comments"
table. This functionality is intended as an automatic annotation identification tool prior to thorough examination of the data in \code{\link{compMS2Explorer}}.
The intention is that automatic annotations can be used in the metID.rtPred
retention prediction function as part of a seamless first-pass workflow.}

\item{possContam}{numeric how many times does a possible annotation have
to appear in the automatically generated possible annotations for it to be
considered a contaminant and therefore not added to the "metID comment" table (default = 3, i.e. if a database name appears more than 3 times in the 
automatic annotation table it will be removed).}

\item{verbose}{logical if TRUE display progress bars.}
}
\description{
The function will automatically run as a parallel computation is the compMS2 object was created in parallel.
}
