% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metID.chemSim.R
\name{metID.chemSim}
\alias{metID.chemSim}
\title{tanimoto chemical similarity first network neighbours}
\usage{
metID.chemSim(object, ...)
}
\arguments{
\item{object}{a "compMS2" class object.}

\item{autoPossId}{logical if TRUE the function will automatically add the name
of the top annotation based on mean maximum 1st neighbour chemical similarity
above the minimum chemical similarity score (default = FALSE). Caution if TRUE
this will overwrite any existing possible_identities in the "metID comments"
table. This functionality is intended as an automatic annotation identification tool prior to thorough examination of the data in \code{\link{compMS2Explorer}}.
The intention is that automatic annotations can be used in the metID.rtPred
retention prediction function as part of a seamless first-pass workflow.}

\item{minSimScore}{numeric must be values between 0-1 minimum tanimoto chemical similarity score (default = 0.8). Any mean maximum 1st neighbour chemical 
similarity scores will be considered for automatic possible identity 
addition to the "metID comments"
 table, the annotation with the highest mean maximum 1st neighbour chemical similarity score will then be automatically added to the "metID comments" table
 in the \code{\link{compMS2Explorer}}.}

\item{possContam}{numeric how many times does a possible annotation have
to appear in the automatically generated possible annotations for it to be
considered a contaminant and therefore not added to the "metID comment" table (default = 3, i.e. if a database name appears more than 3 times in the 
automatic annotation table it will be removed).}

\item{bitsChemFP}{numeric values between 1024-4096 number of most frequent
atom-pairs in the DrugBank database see the ChemmineR function \code{\link{desc2fp}} for more details (default = 1024).}

\item{minEdges}{numeric minimum number of edges (i.e. connected/adjacent nodes/spectra) 
to consider a node/spectrum for chemical similarity (default=2). 
This filtration is performed after removal of isobaric spectra.
For example nodes with only one edge/adjacent node are more likely to produce 
false-positive annotations. For more robust chemical similarity scoring 
consider increasing this number.}

\item{verbose}{logical if TRUE display progress bars.}
}
\description{
tanimoto chemical similarity first network neighbours
}
\details{
this function can only be utilized after running \code{\link{metID.corrNetwork}} and/or \code{\link{metID.specSimNetwork}}. The purpose of this function is to provide first-pass automatic metabolite annotation. The tanimoto chemical similarity score is first calculated from a 1024-4096 bit chemical fingerprint for every best annotation SMILES code. For annotations of each composite spectrum the maximum chemical similarity score with any first neighbours (either by correlation and/or spectral similarity) are identified and the weighted arithmetic mean maximum chemical similarity score of 1st neighbours calculated. The mean is weighted based on the mean spectral similarity and/or
correlation coefficient value for an edge pair (i.e. if two spectra are connected by both spectral similarity and
correlation then a mean value of the two will be calculated and used as the weight).
This is to ensure that the more similar or highly correlated two composite spectra
are the higher the contribution to the maximum chemical similarity score. 
A new column "MMNNCSS" is added to the best annotation tables for any composite spectra with at least one composite spectrum network neighbour. Furthermore, the best Annotation table is
sorted according to this new likely annotation score, this give the user a rapid
means to establish a likely annotation based on chemical similarity with neighbouring node annotations.
 
 Optionally, the top annotations for a composite spectrum can be automatically 
 added to the "metID comments" table in the compMS2Explorer application. If the 
 argument \strong{autoPossId} is TRUE (default = FALSE) the function will automatically add the name
 of the top annotation based on mean maximum 1st neighbour chemical similarity
 above the minimum chemical similarity score (argument minSimScore, default = 0.8).
}
\examples{
library(compMS2Miner)
compMS2Example <- metID(compMS2Example, 'chemSim', minSimScore=0.8, 
                        autoPossId=TRUE)
}
