% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metID.compMS2ToMsp.R
\name{metID.compMS2toMsp}
\alias{metID.compMS2toMsp}
\title{msp file from compMS2 class object}
\usage{
metID.compMS2toMsp(object, studyName = NULL, existingMsp = NULL,
  outputDir = getwd(), onlyCommented = FALSE, ...)
}
\arguments{
\item{object}{a "compMS2" class object}

\item{studyName}{character a study name tag for your dataset.}

\item{existingMsp}{character a full path or web address of an existing msp database
file to concatenate to. The new entries will appear underneath the last entry 
of the existing msp datafile file. If this argument is supplied then this file will be
overwritten.}

\item{outputDir}{character full path to a directory to write the msp file to.
Default is to take the current working directory obtained from \code{\link{getwd}}.}

\item{onlyCommented}{logical (default = FALSE) should only metabolites recorded
in the comments table be included or if true all of the composite spectra.}
}
\value{
the msp file will be written to a file tagged with the study name and date in the output directory.
}
\description{
msp file from compMS2 class object
}
\details{
this function converts a compMS2 class object to an msp database
file inserting experimental details and taking annotation data from the 
comments table and best annotations. This creates a means to database successive
experiments and potentially concatenate to a laboratories existing spectral 
database. In this way previously annotated spectra from other experiments 
can be matched to new datasets. A unique splash code is also generated for
each composite spectrum and added to each database entry in the msp file.
install splash R from GitHub using devtools:
 devtools::install_github("berlinguyinca/spectra-hash", subdir="splashR")
}
