% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metID.corrNetwork.R
\name{metID.corrNetwork}
\alias{metID.corrNetwork}
\title{CompMS2 correlation network generation}
\usage{
metID.corrNetwork(object, ...)
}
\arguments{
\item{object}{A "compMS2" class object.}

\item{peakTable}{a data.frame in the form observation (samples) in columns and
variables (Mass spectral signals) in rows. The first 3 columns must consist of:
\enumerate{
 \item EIC number or unique peak identifier.
 \item mass-to-charge ratio of peak group.
 \item median/ peak apex retention time in seconds. 
 }
These columns are utilized in the final network visualization.}

\item{obsNames}{character vector of observation (i.e. sample/ QC/ Blank) names to identify appropriate observation (sample) columns.}

\item{corrThresh}{correlation coefficient threshold to group features within
a retention time cluster. If no value is supplied the default is to estimate
an optimal cut-off value based on network attributes from a series of correlation cut-off values. This is based upon the relationship between the 
number of nodes, edges, number of clusters and the network density (i.e. n actual edges/n potential edges) at each correlation cut-off value. A plot showing
the result of this estimation will be generated.}

\item{corrMethod}{character correlation method see \code{\link{cor}} for details. default "spearman".}

\item{delta}{numeric maximum p-value (following multiple testing correction) above #' which the null hypothesis (no correlation) is rejected.}

\item{MTC}{character Multiple Testing Correction default is "none", see \code{\link{p.adjust.methods}} for
details of options. ("holm", "hochberg", "hommel", "bonferroni", "BH", "BY", "fdr", "none").
Any p-values after multiple testing correction above the value of delta will have their
corresponding correlation coefficents replaced with zero.}

\item{maxNodes}{numeric above the maximum nodes the function will use the large graphing algorithm of \code{\link{igraph}}. See \code{\link{igraph::with_lgl}} else
the function uses the  Fruchterman-Reingold layout algorithm. See \code{\link{igraph::with_fr}}}

\item{MS2only}{numeric 3 options (1-3) if 1 All EICs above corrThresh returned, if 2 only non-MS2 matched EICs which are first neighbours of at least one MS2-matched MS2 are returned, if 3 only MS2-matched EICs are returned (default = 1).}

\item{minClustSize}{numeric minimum number of connected nodes that is cluster size (default =3). If a cluster
of nodes is less than this number then it will be removed.}
}
\value{
"compMS2" class object with an additional network graph of any peakTable features above the correlation threshold.
}
\description{
Uses the MS1features matched to MS2 data to generate a correlation network to view in compMS2Explorer
}
\references{
Koh Aoki, Yoshiyuki Ogata, and Daisuke Shibata
Approaches for Extracting Practical Information from Gene Co-expression Networks in Plant Biology
Plant Cell Physiol (2007) 48 (3): 381-390 first published online January 23, 2007 doi:10.1093/pcp/pcm013
}
