% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metID.dbAnnotate.R
\name{metID.dbAnnotate}
\alias{metID.dbAnnotate}
\title{Annotate unknown features in a CompMS2 class object to database 
entries based on monoisotopic mass}
\source{
\url{http://fiehnlab.ucdavis.edu/staff/kind/Metabolomics/MS-Adduct-Calculator/}
}
\usage{
metID.dbAnnotate(object, ...)
}
\arguments{
\item{object}{a compMS2 class object.}

\item{ppm}{a ppm error to match MS1 features to data base entries.}

\item{metDB}{a metabolite data base (see default ?HMDB for table format).
Other currently available data bases include DrugBank, T3DB and ReSpect 
databases. Matching using multiple databases is also possible.}

\item{SubStrs}{substructures table containing substructure mass shifts 
(see default ?Substructure_masses for table format)}
}
\value{
a compMS2 class object containing potential metabolite annotations.
}
\description{
unknown metabolite identification. MS1 features within a CompMS2
object are matched against a metabolite database based on ppm error tolerance.
Possible electrospray adducts and substructure mass shifts are calculated for
all data base entries. Substructure mass shifts are supplied in a 
substructure masses table (see default Substructure_masses).
}

