% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/metID.dbProb.R
\name{metID.dbProb}
\alias{metID.dbProb}
\title{selects best annotations based on substructure annotations identified
by \code{\link{subStructure.Annotate}}}
\usage{
metID.dbProb(object, ...)
}
\arguments{
\item{object}{a compMS2 class object}

\item{nameFeat}{a unique name for a composite spectrum of interest. If not
supplied (default) all most probable annotations are decided automatically
and for all composite spectra. Previous most probable annotations will not
be overwritten if the function is run more than once.}

\item{DBids}{unique database identifier for a specific composite spectrum, in
combination with nameFeat argument.}
}
\value{
a compMS2 class object with most probable annotation(s)
}
\description{
Most probable database annotations either automatically
decided based on substructure type detected by the
\code{\link{subStructure.Annotate}} or user supplied most probable annotations
one composite spectrum at a time.
}

