% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metID.metFrag.R
\name{metID.metFrag}
\alias{metID.metFrag}
\title{metFrag \emph{in silico} fragmentation query wrapper function}
\source{
\url{http://c-ruttkies.github.io/MetFrag/projects/metfragcl/} developed based on the command line .jar file (MetFrag2.3-CL.jar) downloaded on 2016/07/12
\enumerate{
\item MetFrag relaunched: incorporating strategies beyond in silico fragmentation: C Ruttkies, E L Schymanski, S Wolf, J Hollender, S Neumann Journal of Cheminformatics 2016 8:3
\item In silico fragmentation for computer assisted identification of metabolite mass spectra: S Wolf, S Schmidt, M Müller-Hannemann, S Neumann BMC bioinformatics 11 (1), 148
}
}
\usage{
metID.metFrag(object, ...)
}
\arguments{
\item{object}{a "compMS2" class object.}

\item{featureSubSet}{character vector of composite spectra names (e.g. CC_1, CC_2 etc.) otherwise the default is to perform metFrag queries on all composite spectra.}

\item{adductTable}{data.frame containing adduct names, MetFrag number codes
and ionization mode. A default data.frame is internal to the package but this
table is fully customizable if future adduct types are added. See View(metFragAdducts)
to see the format of this table. N.B. The metFrag adduct names must perfectly match
those supplied to the \code{\link{adduct2mass}} function during database
annotation using \code{\link{metID.dbAnnotate}}.}

\item{keepTempFiles}{logical default = FALSE, txt, sdf and csv results files will
be created as temporary files otherwise if TRUE files will be retained in named subdirectories (see details).}

\item{maxTreeDepth}{numeric maximum tree depth (default = 2). If 2 then
fragments of fragments are also considered but will increase computation time.}

\item{frag_mzabs}{numeric delta predicted-observed fragment mass accuracy for matching.}

\item{minMetFragScore}{numeric minimum mean total ion current explained and metFrag score (default = 0.9)
the candidate with the highest score above this minimum will be automatically added to the Comments
table. The argument autoPossId must also be set to TRUE.}

\item{autoPossId}{logical if TRUE the function will automatically add the name
of the top annotation based on mean total ion current explained and metFrag score
 (default = FALSE). Caution if TRUE
this will overwrite any existing possible_identities in the "metID comments"
table. This functionality is intended as an automatic annotation identification tool prior to thorough examination of the data in \code{\link{compMS2Explorer}}.
The intention is that automatic annotations can be used in the metID.rtPred
retention prediction function as part of a seamless first-pass workflow.}

\item{possContam}{numeric how many times does a possible annotation have
to appear in the automatically generated possible annotations for it to be
considered a contaminant and therefore not added to the "metID comment" table (default = 3, i.e. if a database name appears more than 3 times in the 
automatic annotation table it will be removed).}

\item{verbose}{logical if TRUE display progress bars.}
}
\value{
a compMS2 class object containing metFrag \emph{in silico} fragmentation 
results which can be visualized in \code{\link{compMS2Explorer}}.
}
\description{
performs metFrag (msbi.ipb-halle.de/MetFrag/) \emph{in silico} combinatorial
fragmentation. Local chemical structure data files (.sdf) are created from 
most probable annotation canonical SMILES codes. Temporary local sdf files 
and metfrag parameter files (.txt) are  created on a composite spectrum by 
composite spectrum basis and
\emph{in silico} fragmentation performed. Results are read back into R and stored in the compMS2 class object as results tables. In addition the temporary sdf,
.txt and .csv metfrag results files may also optionally be kept (keepTempFiles = TRUE) and are saved in a subdirectory structure (see \emph{details}). The function will automatically run as a parallel computation is the compMS2 object was created in parallel. The adduct types and codes which MetFrag command line tool 
will work with are included in a default internal table. See data(metFragAdducts)
 the following electrospray adducts are contained within:
positive mode: '[M+H]+', '[M+NH4]+', '[M+Na]+', '[M+K]+'
negative mode: '[M-H]-', '[M+Cl]-', '[M-H+HCOOH]-', '[M-H+CH3COOH]-' 
All other database annotations of other electrospray adducts will be discarded.
If the format of the metFragAdducts table is correctly followed then additional
MetFrag adduct types can be added see View(metFragAdducts). This ensures that
the adduct types are customizable and can be modified to incorporate future
availability. See argument metFragAdducts below.
}
\details{
if keepTempFiles = FALSE, Results directories are generated in the current working directory: for each MS1 feature matched to MS2 data a results 
                     directory is created named ("MetFrag_results"). Subdirectories are then created within the results directory named after each composite spectrum name.
                     Assuming MetFrag returned any results the following files should appear within the subdirectories:
\enumerate{
\item MetFrag parameters files : MetFrag parameter files (.txt) are saved in each result directory. See \url{http://c-ruttkies.github.io/MetFrag/projects/metfragcl/} for details.
\item localSdf file : the local sdf (chemical structure data files) are saved in 
                    each result directory. This file is used as a local database for MetFrag \emph{in silico} fragmentation.
\item result .csv file: the metFrag results are returned as a comma seperated values text file (.csv).
}
}
