% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metID.optimConsensus.R
\name{metID.optimConsensus}
\alias{metID.optimConsensus}
\title{differential evolution build consensus weight optimization}
\usage{
metID.optimConsensus(object, ...)
}
\arguments{
\item{object}{a "compMS2" class object.}

\item{...}{additional arguments to \code{\link{DEoptim.control}}.}

\item{include}{character vector of 7 options to build consensus combinatorial
metabolite identification see Details below for a description of each. If
specific options are not supplied as a character vector then the default 
is to consider all 7. i.e. 
c('massAccuracy', 'spectralDB', 'inSilico', 'rtPred', 'chemSim', 'pubMed', 'substructure').}

\item{autoPossId}{logical if TRUE the function will automatically add the names
of the top annotation based on mean consensus annotation score to the 
"metID comments" table (default = FALSE). Caution if TRUE this will overwrite 
any existing possible_identities in the "metID comments" table. 
This functionality is intended as an automatic metabolite annotation 
identification tool prior to thorough examination of the data in 
\code{\link{compMS2Explorer}} as part of an objective and seamless 
first-pass annotation workflow. The mean build consensus score can consist
of many orthogonal measurements of metabolite identification and a means
to rapidly rank metabolite annotations.}

\item{minMeanBCscore}{numeric minimum mean consensus score (values between 0-1),
if argument autoPossId is TRUE any metabolite annotations above this value
will be automatically added to the "metID comments" table. (if argument not
supplied the default is the upper interquartile range of the mean BC score).}

\item{possContam}{numeric how many times does a possible annotation have
to appear in the automatically generated possible annotations for it to be
considered a contaminant and therefore not added to the "metID comment" table (default = 3, i.e. if a database name appears more than 3 times in the 
 automatic annotation table it will be removed).}

\item{specDbOnly}{logical if TRUE then only spectra matched to a spectral
database will be considered. These annotations are identified by the flag
"metID.matchSpectralDB" in the comments table. The default FALSE means that
all metabolites in the metID comments table will be considered.}

\item{popSize}{numeric number of population members (see the NP argument in \code{\link{DEoptim.control}}). The default is 10 * length(include).}

\item{itermax}{numeric the maximum number of iterations (population generation)
allowed. (default = 100). See \code{\link{DEoptim.control}} for further details.}

\item{plotInterval}{the number of iterations before plotting the algorithms progress (default=20). Smaller values may slightly slow the process.}
}
\description{
differential evolution build consensus weight optimization
}
\details{
uses the package \code{\link{DEoptim}} to calculate the optimum
weighting of all included consensus scores to accurately rank the known
annotations (taken from the "metID comments" table in \code{\link{compMS2Explorer}}). These global parameters can then be used to rank the unknown annotations of other
unannotated composite spectra based on the optimum weighted mean consensus score.
The annotations above a certain score (minMeanBCscore) can also be automatically
added to the "metID comments" table. The ongoing differential evolution process will appear in a plot window with a loess fit line in red highlighting any reduction
in the mean rank of the training set annotations (from "metID comments" table) as
the genetic process evolves. This metaheuristic global optimization process
can help to maximise the parameters for accurate metabolite annotation and ranking.
}
\examples{
compMS2Example <- metID(compMS2Example, 'optimConsensus')
}
\references{
\enumerate{
\item David Ardia, Katharine M. Mullen, Brian G. Peterson, Joshua Ulrich (2015).
'DEoptim': Differential Evolution in 'R'. version 2.2-3.

\item Katharine Mullen, David Ardia, David Gil, Donald Windover, James Cline
(2011). 'DEoptim': An R Package for Global Optimization by Differential
Evolution. Journal of Statistical Software, 40(6), 1-26. URL
http://www.jstatsoft.org/v40/i06/.

\item Ardia, D., Boudt, K., Carl, P., Mullen, K.M., Peterson, B.G. (2010).
Differential Evolution with 'DEoptim': An Application to Non-Convex Portfolio
Optimization. The R Journal, 3(1), 27-34. URL
http://journal.r-project.org/archive/2011-1/2011-1_index.html.

\item Ardia, D., Ospina Arango, N., Giraldo Gomez, N. (2010). Jump-Diffusion
Calibration using Differential Evolution. Wilmott Magazine, Issue 55
(September), 76-79. URL http://www.wilmott.com/.

 \item Kenneth V. Price, Rainer M. Storn and Jouni A. Lampinen (2006). 
 Differential Evolution - A Practical Approach to Global Optimization. Berlin Heidelberg:
  Springer-Verlag. ISBN 3540209506. 
  }
}
