% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metID.predSMILES.R
\name{metID.predSMILES}
\alias{metID.predSMILES}
\title{Phase II metabolite prediction}
\usage{
metID.predSMILES(object, ...)
}
\arguments{
\item{object}{a compMS2 object (only possible when probable annotations
i.e. metID.dbProb has been already performed).}
}
\value{
a compMS2 class object containing predicted Phase II metabolites from
most probable database annotations.
}
\description{
calculates possible Phase II metabolite structures from
canonical SMILES codes of most probable metabolite annotations. Currently
the algorithm predicts only certain possible Phase II metabolites and no
Phase I metabolism. The simple cases of acyl-, hydroxyl- and amine- sulfates
and glucuronides and glycine conjugates are predicted based on the presence
of these functional groups within the SMILES code.
}
