% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metID.specSimNetwork.R
\name{metID.specSimNetwork}
\alias{metID.specSimNetwork}
\title{CompMS2 spectral similarity network generation}
\usage{
metID.specSimNetwork(object, ...)
}
\arguments{
\item{object}{A "compMS2" class object.}

\item{minDotProdThresh}{minimum dot product spectral similarity score. 
If no value is supplied the default is to estimate
an optimal cut-off value based on network attributes from a series of spectral similarity cut-off values. This is based upon the relationship between the 
number of nodes, edges, number of clusters and the network density (i.e. n actual edges/n potential edges) at each spectral similarity cut-off value. A plot showing
the result of this estimation will be generated.}

\item{binSizeMS2}{numeric MS2 bin size for spectral similarity matching (default = 0.1)}

\item{minClustSize}{numeric minimum number of connected nodes that is cluster size (default =3). If a cluster
of nodes is less than this number then it will be removed.}
}
\value{
"compMS2" class object with an additional network graph of any peakTable features above the correlation threshold.
}
\description{
generates a dot product spectral similarity network from the MS/MS fragmentation data. The resulting spectral similarity network can then be viewed in compMS2Explorer. Utilizes the \code{\link{graph}} function of the \code{\link{igraph}} package.
}
\references{
Koh Aoki, Yoshiyuki Ogata, and Daisuke Shibata
Approaches for Extracting Practical Information from Gene Co-expression Networks in Plant Biology
Plant Cell Physiol (2007) 48 (3): 381-390 first published online January 23, 2007 doi:10.1093/pcp/pcm013
}
