% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mfSearchPUG.R
\name{mfSearchPUG}
\alias{mfSearchPUG}
\title{Search pubmed compound for molecular formula using the pubchem power user gateway (PUG)}
\usage{
mfSearchPUG(mf = "C10H21N")
}
\arguments{
\item{mf}{molecular formula character vector of length one e.g. 'C10H21N'.}
}
\value{
returns a character vector of pubmed compound cids matching the molecular
formula
}
\description{
Search pubmed compound for molecular formula using the pubchem power user gateway (PUG)
}

