% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monoMassMatch.R
\name{monoMassMatch}
\alias{monoMassMatch}
\title{Monoisotopic mass matching unknown to data base entry}
\usage{
monoMassMatch(unknowns = NULL, metMasses.df = NULL, esiAdducts = NULL,
  subStrMasses = NULL, mode = "pos", ppm = 10, nCores = NULL)
}
\description{
Monoisotopic mass matching unknown to data base entry
}
