% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/MS2fileInfo.R
\name{MS2fileInfo}
\alias{MS2fileInfo}
\title{MS2 file information}
\usage{
MS2fileInfo(MS2file = NULL, TICfilter = NULL)
}
\arguments{
\item{MS2file}{MS2 file imported into R via the readMzXml package as a list}
}
\value{
data frame number of rows equal to the number of scans and 6
observations:
1. "MS.scanType" = MS scan type, 1 or 2.
2. "precursorMz" = numeric mass-to-charge ratio for each MS2 scan
3. "retentionTime" = numeric precusor retention time for each MS2 scan
4. "TIC" = Total ion current for this scan.
5. "TICaboveFilter" = boolean if TIC above minimum TIC filter equal 1, else 0
6. "precursorIntensity" = MS2 precursor intensity, if MS1 scan returns
zero
7. "collisionEnergy" = collision energy (eV)
8. "basePeakMz" = mass-to-charge ratio of the base peak for the scan
9. "basePeakIntensity" = intensity of the base peak for the scan
10. "precursorScanNum" = precursor scan number (MS2) or scan number (MS1)
}
\description{
Extract precursor mass-to-charge ratio, retention time, scan type and total
ion current for each MS2 file scan.
}

