% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimCutOff.R
\name{optimCutOff}
\alias{optimCutOff}
\title{optimum cutoff for a correlation or spectral similarity matrix}
\usage{
optimCutOff(x = NULL, cutOffSeq = seq(0.01, 1, 0.01), diffConsecVals = 1.5
  * 10^-3, maxCutOff = 0.95)
}
\arguments{
\item{x}{matrix of correlation coefficients or spectral similarity values
column and row names must match}

\item{cutOffSeq}{numeric a vector of cut-off values to test. (default=seq(0.01, 1, 0.01) a numeric vector of length 100)}

\item{diffConsecVals}{numeric the scaled difference between consecutive values
to identify the plateau in the network density (default=1.5*10^-3 or 0.0015).}
}
\value{
a list containing two named elements "estCutOff" a numeric estimated optimal cut-off value and "testData" a matrix
of the test results at each cut-off value. The function also plots the result.
}
\description{
optimum cutoff for a correlation or spectral similarity matrix
}
\references{
Koh Aoki, Yoshiyuki Ogata, and Daisuke Shibata
Approaches for Extracting Practical Information from Gene Co-expression Networks in Plant Biology
Plant Cell Physiol (2007) 48 (3): 381-390 first published online January 23, 2007 doi:10.1093/pcp/pcm013
}
