% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PMIDsearch.R
\name{PMIDsearch}
\alias{PMIDsearch}
\title{customized PMID search function adapted from PubChemWordcloud package v 0.3.2}
\usage{
PMIDsearch(keys = NULL, n = 1000)
}
\arguments{
\item{keys}{character vector of compound names to search pubmed with}

\item{n}{numeric maximum number of results to return}
}
\description{
customized PMID search function adapted from PubChemWordcloud package v 0.3.2
}
