% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/publishApp.R
\name{publishApp}
\alias{publishApp}
\title{Publish your compMS2explorer app on shinyapps.io}
\usage{
publishApp(object, appName = NULL, writeDir = NULL, ...)
}
\arguments{
\item{object}{a compMS2 class object}

\item{appName}{character name for your new app (e.g. 'compMS2example').}

\item{writeDir}{character full path to a directory to save the compMS2 results
and the current shiny app to. This zip file can then be shared with others.}

\item{...}{further arguments to the \code{\link{deployApp}} function}
}
\description{
you must hold a shinyapps.io account and have a token.
use this command from the package rsconnect to set up your token
\code{setAccountInfo(name='your_shinyapps.io_user_name', token='your_token', secret='your_secret')}
Following this the app can be deployed on your account
}

