% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pubMedSearch.R
\name{pubMedSearch}
\alias{pubMedSearch}
\title{return cleaned abstracts from pubmed from searched key words}
\usage{
pubMedSearch(keys = NULL, n = 500, maxChar = 50, ...)
}
\arguments{
\item{keys}{character vector of compound names to search pubmed with}

\item{n}{numeric maximum number of results to return. The maximum and default is 500.}

\item{maxChar}{numeric maximum number of characters in cleaned abstract words to return.}

\item{...}{further arguments to the \code{\link{cleanAbstracts}} function.}
}
\value{
a list containing 3 named elements:
1. titles character vector of Abstract title(s)
2. abs character vector of abstract text(s).
3. clAbs clean abstract word frequency data.frame with column names 'word' and 'freq'.
}
\description{
return cleaned abstracts from pubmed from searched key words
}
\seealso{
\code{\link{PubMedWordcloud}}, \code{\link{getAbstracts}}, \code{\link{cleanAbstracts}}.
}
