% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ReSpect.R
\docType{data}
\name{ReSpect}
\alias{ReSpect}
\title{ReSpect Database for Phytochemicals}
\format{A data frame with 3078 rows and 8 variables}
\source{
\url{http://spectra.psc.riken.jp/menta.cgi/respect/download/download}
}
\usage{
data(ReSpect)
}
\description{
This dataset contains fields from the ReSpect database for Phytochemicals
The variables are as follows:
}
\details{
\itemize{
 \item WebAddress web address for database (when combined with the
 Unique_DB_ID entry this will provide a direct link to the online database
 entry e.g. Web address : "www.hmdb.ca/metabolites/" +
 Unique_DB_ID : "HMDB00001" = \url{www.hmdb.ca/metabolites/HMDB00001})
 \item Unique_DB_ID. ReSpect unique reference number (PM000101 -- PT211770)
 \item name. ReSpect entry name
 \item monoisotopic_weight. Monoisotopic weight of ReSpect entry
 (61.05 -- 2804.23)
 \item SMILES. Canonical SMILES code of ReSpect entry
 \item molecular_formula. Molecular formula of ReSpect entry
 \item cas_number. CAS registry number of ReSpect entry
 \item species_id. if available plant species name
 }
}
\keyword{datasets}

