% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runGitHubApp.R
\name{runGitHubApp}
\alias{runGitHubApp}
\title{run github shiny modified from shiny and devtools}
\usage{
runGitHubApp(repo = NULL, subdir = NULL, auth_token = NULL)
}
\arguments{
\item{repo}{character github username and repository name. in the form "username/repositoryName"}

\item{subdir}{character sub-directory of the repo containing the shiny and data.}

\item{auth_token}{character private repo authorization token.}
}
\description{
run github shiny modified from shiny and devtools
}

