% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signalGrouping.R
\name{signalGrouping}
\alias{signalGrouping}
\title{Signal grouping}
\usage{
signalGrouping(spectrum.df = NULL, mzError = 0.001, minPeaks = 5)
}
\arguments{
\item{spectrum.df}{a dataframe or matrix with two or more columns:
1. Mass/ Mass-to-charge ratio
2. Intensity}

\item{mzError}{interpeak absolute m/z error for signal grouping 
(Default = 0.001)}
}
\value{
dataframe of m/z grouped signals, the m/z values of the input 
dataframe/ matrix peak groups are averaged and the signal intensities summed.
}
\description{
Euclidean distances between m/z signals are hierarchically clustering using 
the average method and the composite spectrum groups determined by a absolute
error cutoff
}

