% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smiles2Form.R
\name{smiles2Form}
\alias{smiles2Form}
\title{Convert SMILES code to atomic formula}
\usage{
smiles2Form(SMILES = NULL)
}
\arguments{
\item{SMILES}{character vector of SMILES codes to convert}
}
\value{
a character vector the formula(e).
}
\description{
Convert SMILES code to atomic formula
}
