% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smiles2MonoMassForm.R
\name{smiles2MonoMassForm}
\alias{smiles2MonoMassForm}
\title{Convert SMILES code to monoisotopic mass and formula}
\usage{
smiles2MonoMassForm(SMILES = NULL)
}
\arguments{
\item{SMILES}{character vector of SMILES codes to convert}
}
\value{
a named numeric vector of same length as the SMILES input containing the
monoisotopic mass(es) and named using the formula(e).
}
\description{
Convert SMILES code to monoisotopic mass and formula
}
