% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subFormulae.R
\name{subFormulae}
\alias{subFormulae}
\title{subtract atomic formula y from atomic formula x}
\usage{
subFormulae(x = NULL, y = NULL)
}
\arguments{
\item{x}{character vector of atomic formulae (must be same length as y).}

\item{y}{character vector of atomic formulae (must be same length as x).}
}
\description{
subtract atomic formula y from atomic formula x
}
