% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subsetCompMS2.R
\name{subsetCompMS2}
\alias{subsetCompMS2}
\title{subset compMS2 class object using a vector of spectra names}
\usage{
subsetCompMS2(object, ...)
}
\arguments{
\item{object}{a "compMS2" class object.}

\item{specNames}{character vector of composite spectrum names.}

\item{corrNetworkNodes}{logical should all the members of correlation network
cluster be returned? 
(default = FALSE).}

\item{rtGroups}{logical should all the nodes within the same retention time
group also be returned? (default = FALSE).}

\item{onlyAdductId}{logical should only ESI adduct/in-source fragments which
have been identified and are connected to the specNames argument be returned?
(default = FALSE). If a features of the specNames argument has no adducts
identified then it will be returned in isolation.}

\item{nodeOrder}{neighbourhood order (see \link{ego} function from igraph package)
for details. If argument "onlyAdductID" is TRUE then only nodes less
than order away from each of the specNames argument will be returned. This
occurs after onlyAdductId subset features have been identified. This is performed
to reduce complexity of very dense network clusters (default = 2).}
}
\value{
a "compMS2" class object with the composite spectra and all metID 
information removed.
}
\description{
subset compMS2 class object using a vector of spectra names
}
