% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subStructure.Annotate.R
\name{subStructure.Annotate}
\alias{subStructure.Annotate}
\title{composite spectra substructure annotation}
\usage{
subStructure.Annotate(object, ...)
}
\arguments{
\item{Frag_mzabs}{Absolute mass accuracy difference to identify neutral losses 
and fragments in composite spectra (default = 0.01).}

\item{SubStrs}{substructure data frame (default = Substructure_masses)
see ?Substructure_masses for details of the mandatory table fields/ format}

\item{minRelInt}{minimum relative intensity to consider a spectral signal
for substructure annotation (default = 5 i.e. 5\% rel. int.).}
}
\description{
composite spectra substructure annotation
}
