% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subStructure.prob.R
\name{subStructure.prob}
\alias{subStructure.prob}
\title{Identifies probable substructure type}
\usage{
subStructure.prob(object, ...)
}
\arguments{
\item{a}{compMS2 object}
}
\value{
a data.frame of probable substructure annotations for each composite
spectrum, ranked by the sum of the relative intensities for that substructure
type
}
\description{
Identifies probable substructure type based on the summed
relative intensites and therefore proportion of total composite spectrum 
intensity explained.
}

