% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subStructure.R
\name{subStructure}
\alias{subStructure}
\title{Identify substructures within composite spectra}
\usage{
subStructure(object, ...)
}
\arguments{
\item{...}{option arguments to be passed along.}

\item{object.}{a compMS2 class object obtained from the function CompMSset}

\item{method.}{"Annotate" annotation of possible substructure neutral losses/
fragments in composite spectra, "prob" identify most probable substructure
identification for a composite spectra and "probSummary" summary of probable
substructure annotations for each composite spectrum.}
}
\value{
A compMS2 object with substructure annotated composite spectra.
}
\description{
Identify substructures within composite spectra
}
\seealso{
\link{subStructure.Annotate}, \link{subStructure.prob}, \link{subStructure.probSummary}
}
