% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trueFalseSum.R
\name{trueFalseSum}
\alias{trueFalseSum}
\title{summary of true positive and false negative assignments for each score metric}
\usage{
trueFalseSum(object, ...)
}
\arguments{
\item{object}{a "compMS2" class object}

\item{n}{integer the number of top annotations to consider (default = 5).
i.e. if the correct structure is found in the top 5 structures then it is
considered a true positive.}

\item{minSimScore}{numeric chemical similarity score (values between 0 and 1).
This score is used to identify the compound contained in the metID comments
table (default = 1). The is performed rather than by name as there can
be multiple hits of the same structure in the top annotations. To check
if the metric is in roughly the right ballpark highly similar structures to 
the correct annotations can be considered by setting the argument to a lower 
Tanimoto chemical similarity score (e.g. 0.9).}

\item{specDbOnly}{logical if TRUE (default) then only the spectral database
annotations are used to calculate the true positive and false negative rates. If FALSE
then all possible_identities in the metID comments table are used.}

\item{verbose}{logical if TRUE display progress bars.}
}
\value{
a list summarizing the true positive and false negative outputs, a 
summary plot of the results and also weighted average rank for each score metric.
}
\description{
summary of true positive and false negative assignments for each score metric
}
\details{
The weighted average takes in to account the number of possible candidates
by weighting the mean correct rank by the number of possible candidates. e.g. In this
way a rank of 1 out of 10 possible structures will be ranked lower than a rank of 3 out
of 1,000 possible structures. The latter is a more important/impressive reflection of 
ranking ability of than the former. If a more extensive chemical database is
utilized such as PubMed compound for example then a far larger number of
candidates may be considered than a much smaller curated database such as HMDB.
}
