% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/combineMS2.R
\name{combineMS2}
\alias{combineMS2}
\title{combineMS2 ions either within and/ or between composite spectra}
\usage{
combineMS2(object, ...)
}
\arguments{
\item{...}{option arguments to be passed along.}

\item{object.}{a compMS2 class object obtained from the function CompMSset}

\item{method.}{"Ions" intra-spectrum ions grouping or "Spectra" inter-spectra
ion grouping, composite spectra from multiple MS2 files matched to the same
MS1 feature are matched.}
}
\value{
A compMS2 object with intra-composite spectrum or inter-composite
spectra, grouped, signal summed and group mass-to-charge averaged spectra.
}
\description{
combineMS2 ions either within and/ or between composite spectra
}
\seealso{
\code{\link{combineMS2.Ions}}, \code{\link{combineMS2.Spectra}}
}

