% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/combineMS2.Spectra.R
\name{combineMS2.Spectra}
\alias{combineMS2.Spectra}
\title{combine ions across spectra matched to the same MS1 feature}
\usage{
combineMS2.Spectra(object, ...)
}
\arguments{
\item{mzError}{interpeak absolute m/z error for composite spectra signal grouping}

\item{minPeaks}{Minimum number of peaks per composite spectrum}
}
\description{
combine ions across spectra matched to the same MS1 feature
}

