% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compMS2.R
\name{compMS2}
\alias{compMS2}
\title{Constructor for compMS2 class object from a peak table and MS2 mzXML file(s)}
\usage{
compMS2(MS1features = NULL, mzXMLdir = NULL, nCores = NULL,
  argsCorrNetwork = list(obsNames = NULL, corrThresh = 0.6, corrMethod =
  "spearman", delta = 0.05, MTC = "BH"), mode = "pos", precursorPpm = 10,
  ret = 10, TICfilter = 10000)
}
\arguments{
\item{MS1features}{either a data.frame, full file path as a character string to a  .csv file of a MS1 feature table in the form observation (samples) in columns and
variables (Mass spectral signals) in rows, the first 3 columns must consist of:
\enumerate{
 \item EIC number or unique peak identifier.
 \item mass-to-charge ratio of peak group.
 \item median/ peak apex retention time in seconds. 
 }
 If argument is not supplied a GUI (tcltk) file selection window will open and a .csv file can then be selected.}

\item{nCores}{numeric Number of cores for parallel computation.}

\item{mode}{character Ionisation polarity must be either 'pos' or 'neg'.}

\item{precursorPpm}{numeric Parts per million mass accuracy to match MS1 features to MS2 spectra (ppm)}

\item{ret}{numeric retention time tolerance to match MS1 features to MS2 spectra (+/- seconds).}

\item{TICfilter}{numeric Minimum Total Ion Current to consider an MS2 spectrum. Any MS2 scan
below this threshold will not be considered.}
}
\value{
A compMS2 object
}
\description{
Matches MS1 features to MS2 spectra (.mzXML) files based on a 
mass-to-charge and retention time tolerance. Composite spectra and other data
can subsequently be visualized during any stage of the CompMS2miner
processing workflow. Composite spectra can be denoised, ion signals grouped 
and summed, substructure groups identified, common Phase II metabolites
predicted and features matched to data bases monoisotopic mass data 
and insilico MS2 fragmentation data.
The resulting data can then be readily curated by sending to a local or online
couchDB database.
}

