% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compMS2Create.R
\name{compMS2Create}
\alias{compMS2Create}
\title{create a compMS2 object}
\usage{
compMS2Create(MS2file = NULL, MS1features = NULL, TICfilter = 10000,
  precursorPpm = 10, ret = 10, adducts = FALSE, isoWid = 4)
}
\arguments{
\item{MSfiles}{character vector of mzXML file locations}
}
\description{
create a compMS2 object
}
