% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compMS2explorer.R
\name{compMS2explorer}
\alias{compMS2explorer}
\title{Visualize your CompMS2miner results output using a shiny app.}
\usage{
compMS2explorer(object, ...)
}
\arguments{
\item{object}{a compMS2 class object}

\item{browserLaunch}{logical launch app in web browser (default = TRUE).}
}
\description{
Visualize your CompMS2miner results output using a shiny app.
}

