% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metID.corrNetwork.R
\name{metID.corrNetwork}
\alias{metID.corrNetwork}
\title{CompMS2 correlation network generation}
\usage{
metID.corrNetwork(object, ...)
}
\arguments{
\item{object}{A "CompMS2" class object.}

\item{peakTable}{a data.frame in the form observation (samples) in columns and
variables (Mass spectral signals) in rows.}

\item{obsNames}{character vector of observation (i.e. sample/ QC/ Blank) names to identify appropriate observation (sample) columns.}

\item{corrThresh}{correlation coefficient threshold to group features within
a retention time cluster.}

\item{corrMethod}{character correlation method see \code{\link{cor}} for details. default "spearman".}

\item{delta}{numeric maximum p-value (following multiple testing correction) above 
which the null hypothesis (no correlation) is rejected.}

\item{MTC}{character Multiple Testing Correction default is "BH", see \code{\link{p.adjust.methods}} for
details of options. ("holm", "hochberg", "hommel", "bonferroni", "BH", "BY", "fdr", "none").
Any p-values after multiple testing correction above the value of delta will have their
corresponding correlation coefficents replaced with zero.}

\item{maxNodes}{numeric above the maximum nodes the function will use the large graphing algorithm of \code{\link{igraph}}. See \code{\link{igraph::with_lgl}} else
the function uses the  Fruchterman-Reingold layout algorithm. See \code{\link{igraph::with_fr}}}
}
\value{
"CompMS2" class object with an additional network graph of any peakTable features above the correlation threshold.
}
\description{
Uses the MS1features matched to MS2 data to generate a correlation network to view in compMS2explorer
}

