% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metID.dbAnnotate.R
\name{metID.dbAnnotate}
\alias{metID.dbAnnotate}
\title{Annotate unknown features in a compMS2 class object to database 
entries based on monoisotopic mass}
\usage{
metID.dbAnnotate(object, ...)
}
\arguments{
\item{object}{a compMS2 class object.}

\item{ppm}{numeric mass accuracy error (ppm) to match MS1 features to data 
base entries.}

\item{esiAdducts}{data.frame or character vector of custom electrospray 
adducts can be supplied as a character vector of electrospray adducts or as
a data.frame output from the function \link{adduct2mass}. e.g. 
c("[M-H-NH3-CO-COCH2-C4H6O]-", "[4M-H+Cl]2-", "[2M-H]-", "[3M-3H+Fe2+]-", 
"[M-H-CH2O]-", "[3M-2H]2-", "[M-H-CO2-C3H6]-", "[M-H+CH3COOH]-", "[3M-H]-", 
"[M-2H]2-")}

\item{metDB}{a metabolite data base (see default ?HMDB for table format).
Other currently available data bases include LMSD, DrugBank, T3DB and ReSpect 
databases. Matching using multiple databases is also possible.}

\item{SubStrs}{named numeric vector of substructure mass shifts to consider. 
If argument is not supplied then no substructure mass shifts will be considered.
If the character "All" is supplied then all the substructure mass shifts 
contained in the internal table will be considered 
(see default \link{Substructure_masses} for more information).}

\item{featureSubSet}{character vector of composite spectra names (e.g. CC_1, CC_2 etc.) otherwise the default is to perform database annotation on all composite spectra.}

\item{includeElements}{character vector of element symbols (case-sensitive) 
to include in the metDB argument. Any structure containing an element which 
is not in this inclusion list will not be considered. see ?exactMassEle for 
the internal element table.}

\item{mixtures}{logical (default = FALSE), should mixtures be considered. Any
SMILES codes consisting of multiple non-covalently linked structures will not
be considered (e.g. salts)}

\item{MS1adducts}{logical (default = FALSE), should the adducts included in
the 4th column of the MS1feature table be used to reduce false positive 
assignments. The adducts identified will be used to guide the annotation process,
only the unique adducts/fragments will be used to calculate the expected
mass shift values. If a spectrum has no adduct identified then all of the
unique adducts identified in the dataset will be used.}
}
\value{
a compMS2 class object containing potential metabolite annotations.
}
\description{
unknown metabolite identification. MS1 features within a compMS2
object are matched against a metabolite database based on a mass accuracy
tolerance (ppm).
Possible electrospray adducts/ in-source fragments and substructure mass 
shifts are calculated for all data base entries. Substructure mass shifts 
are supplied as a named numeric vector or default is all the mass shifts in 
the internal substructure masses table (see default Substructure_masses).
Custom ESI adduct/ in-source fragments can be supplied as a character string
of names or the output of the function \link{adduct2mass}. See examples.
}
\examples{
SubStrMassShift <- c(42.010565, 119.004101, 176.03209, 255.988909, 305.068159, 
                     57.021464, 161.014666, 79.956817)
names(SubStrMassShift) <- c("acetyl", "cysteine", "glucuronide", 
                            "glucuronide sulfate", "glutathione", "glycine",
                            "mercapturate", "sulfate")
# custom ESI adducts (default is to consider all ESI adducts/ in-source 
# fragments from supplementary material Beyond Profiling manuscript 
# see references for details)
# The function adduct2mass can interpret ESI adduct names and generate a 
# data.frame of expected mass shifts
customEsiAdducts <- c("[M-H-NH3-CO-COCH2-C4H6O]-", "[4M-H+Cl]2-", "[2M-H]-", 
                      "[3M-3H+Fe2+]-", "[M-H-CH2O]-", "[3M-2H]2-", 
                      "[M-H-CO2-C3H6]-", "[M-H+CH3COOH]-", "[3M-H]-", 
                      "[M-2H]2-")

compMS2Example <- metID(compMS2Example, "dbAnnotate", SubStrs=SubStrMassShift, 
                        esiAdducts=customEsiAdducts)
                                      
}
\references{
\enumerate{
\item Stanstrup, J., Gerlich, M., Dragsted, L.O. et al. 
Anal Bioanal Chem (2013) 405: 5037. doi:10.1007/s00216-013-6954-6
}
}
\seealso{
\link{adduct2mass}.
}
