% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metID.dbProb.R
\name{metID.dbProb}
\alias{metID.dbProb}
\title{selects best annotations based on substructure annotations identified
by \code{\link{subStructure.Annotate}}}
\usage{
metID.dbProb(object, ...)
}
\arguments{
\item{object}{a compMS2 class object}

\item{nameFeat}{character of a unique name of a single composite spectra of
interest. If not supplied (default) all most probable annotations are decided automatically
and for all composite spectra. Previous most probable annotations will not
be overwritten if the function is run more than once.}

\item{DBids}{unique database identifier for a specific composite spectrum, in
combination with nameFeat argument.}

\item{minTimesId}{numeric (default = 2) the minimum number of times a particular
substructure type must be identified for it to be considered. This helps to
limit consideration of neutral losses/fragments that have been identified
once for example by chance.}

\item{verbose}{logical if TRUE display progress bars.}
}
\value{
a compMS2 class object with most probable annotation(s)
}
\description{
Most probable database annotations either automatically 
decided based on substructure type detected by the 
\code{\link{subStructure.Annotate}} or user supplied most probable annotations 
one composite spectrum at a time. Additionally any substructure either 
neutral loss or product ion with an available SMILES code will be matched
against all of the available annotations. The function utilizes the 
\code{\link{cmp.similarity}} function set to mode 2 (using the size of the 
descriptor intersection over the size of the smaller descriptor, to deal
with compounds that vary alot in size) of the ChemmineR package to calculate
the similarity of the substructure to the annotated structure. The average 
score between 0-1 of all of the substructures annotated by the 
\code{\link{subStructure.Annotate}}
function is returned in a new column in the
"Best Annotations" panel in the \code{\link{compMS2Explorer}} and the annotations ranked accordingly. Additionally, if 
either a database annotation corresponds to a substructure type annotated 
(e.g. glucuronide) or is the name of the database entry contains the substructure
name (case-insensitive) then this will be give a maximum top score of 1.
}
\examples{
compMS2Example <- metID(compMS2Example, 'dbProb')
}
\seealso{
\code{\link{subStructure.Annotate}}, \code{\link{metID.dbAnnotate}},
\code{\link{cmp.similarity}}.
}
