% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metID.matchSpectralDB.R
\name{metID.matchSpectralDB}
\alias{metID.matchSpectralDB}
\title{match spectra to spectral databases in the .msp file format (NIST)}
\usage{
metID.matchSpectralDB(object, ...)
}
\arguments{
\item{object}{A "compMS2" class object.}

\item{mspFile}{character string to an online or local .msp spectrum database file.
(default = NULL, dependendent on polarity either a positive or negative mode
version of massbank hosted on github will be downloaded) See \url{https://raw.githubusercontent.com/WMBEdmands/mspFiles/master/MassBank_MSMS_Pos_Rev173_vs1.msp} and \url{https://raw.githubusercontent.com/WMBEdmands/mspFiles/master/MassBank_MSMS_Neg_Rev173_vs1.msp}. Additional .msp files can be downloaded
from the massbank of North America website \url{http://mona.fiehnlab.ucdavis.edu/downloads}.}

\item{minDBDotProdThresh}{minimum dot product spectral similarity score (default = 0.65).}

\item{ppmMS1}{numeric minimum mass accuracy (ppm) between database (.msp)
precursor masses and the CompMS2 composite spectrum MS1 m/z. 
Dot products will only be calculated for database entries within this mass 
accuracy threshold (default=10ppm).}

\item{binSizeMS2}{numeric bin size for between database (.msp) fragment masses and the CompMS2 composite spectrum fragment ions (default = 0.1).}

\item{autoPossId}{logical if TRUE and if the .msp file database entries contain
at the very least a SMILES code (for downstream metabolite identification purposes) 
all msp file database entries will be added to the best annotations table.
and the function will also automatically add the name of the best annotation 
(highest dot product score above the user defined threshold) 
to the comments table and a note ("metID.matchSpectralDB") added to the 
comments column. This represents a form of automated metabolite identification
which can still be modified in the comments table upon visualization with
compMS2Explorer.}

\item{verbose}{logical if TRUE display progress bars.}
}
\value{
"compMS2" class object with any database matches above the minimum dot product score.
}
\description{
match spectra to spectral databases in the .msp file format (NIST)
}
\details{
If the msp file contains smiles codes then any matches above the min
dot prod or minimum proportion of the spectrum explained will be added to the 
best annotations table.
}
