% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metID.matchSpectralDB.R
\name{metID.matchSpectralDB}
\alias{metID.matchSpectralDB}
\title{match spectra to spectral databases in the .msp file format (NIST)}
\usage{
metID.matchSpectralDB(object, ...)
}
\arguments{
\item{object}{A "CompMS2" class object.}

\item{mspFile}{character string to an online or local .msp spectrum database file.
(default = "http://prime.psc.riken.jp/Metabolomics_Software/MS-DIAL/LipidBlast_Posi_Plasma_vs2.msp")}

\item{minDBDotProdThresh}{minimum dot product spectral similarity score (default = 0.8).}

\item{deltaMassAbsMS1}{numeric minimum delta mass accuracy between database (.msp)
precursor masses and the CompMS2 composite spectrum MS1 m/z. 
Dot products will only be calculated for database entries within this mass 
accuracy threshold (default=0.01).}

\item{deltaMassAbsMS2}{numeric bin size for between database (.msp) fragment masses and the CompMS2 composite spectrum fragment ions (default = 0.1).}
}
\value{
"CompMS2" class object with any database matches above the minimum dot product score.
}
\description{
match spectra to spectral databases in the .msp file format (NIST)
}

