% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metID.metFrag.R
\name{metID.metFrag}
\alias{metID.metFrag}
\title{metFrag query function}
\usage{
metID.metFrag(object, ...)
}
\arguments{
\item{object}{a compMS2 class object}

\item{featureSubSet}{optional character vector of feature names to on which to
perform metFrag queries
otherwise the default is to perform metFrag queries on all features.}

\item{metFragJar}{file path to metfrag jar file, if null the external MetFragJarFile.jar
found in the ext data directory will be utilized. Although the maintainers
of CompMS2miner will endeavor to install any updated jar files the latest
version can be downloaded here: 
 \url{https://github.com/c-ruttkies/Tools/raw/master/MetFragCommandLineTool.jar}
 and the full path to the downloaded .jar file can alternatively be supplied.}

\item{keepTempSdf}{logical default = FALSE, sdf, mf and results files will
be created temporarily otherwise temporary files will be retained in named
subdirectories (see details).}
}
\value{
a compMS2 class object containing metFrag insilico fragmentation 
results.
}
\description{
performs metFrag (msbi.ipb-halle.de/MetFrag/) insilico combinatorial
fragmentation. Local chemical structure data files (.sdf) are created from 
most probable annotation canonical SMILES codes. Temporary local sdf files 
and metfrag query files (.mf) are  created on a composite spectrum by 
composite spectrum basis and
insilico fragmentation performed. Results are read back into R and stored in
the compMS2 class object as results tables. In addition the temporary sdf,
mf and metfrag results files may also optionally be kept (keepTempSdf = TRUE)
and are saved in a subdirectory structure.
}
\details{
Results directories: for each MS1 feature matched to MS2 data a results 
                     directory is created. Features and their corresponding 
                     directories are named according to the schema ("EIC_",
                     MS1 feature EIC/unique ID number,"_M",mass-to-charge
                     ratio,"RT",retention time in seconds).

MetFrag query files : MetFrag query files (.mf) are saved in each result
                      directory. 
                See http://c-ruttkies.github.io/MetFrag/projects/commandline/
                for details

Results sdf files : local sdf (chemical structure data files) are saved in 
                    each result directory. 
                    These files are used for MetFrag insilico fragmentation.
}

