% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/monoMassMatch.R
\name{monoMassMatch}
\alias{monoMassMatch}
\title{Monoisotopic mass matching unknown to data base entry}
\usage{
monoMassMatch(unknowns = NULL, metMasses.df = NULL, metMasses = NULL,
  DB_unique_IDs = NULL, DB_entry_names = NULL, DB_SMILES = NULL,
  dbEntryWebAddress = NULL, subStrMasses = NULL, subStrNames = NULL,
  nCores = NULL, mode = "pos", ppm = 10)
}
\description{
Monoisotopic mass matching unknown to data base entry
}

