% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MS1MatchSpectra.R
\name{MS1MatchSpectra}
\alias{MS1MatchSpectra}
\title{MS1featureMatch}
\usage{
MS1MatchSpectra(file.info.df = NULL, MS2file = NULL, mz = NULL,
  RT = NULL, EIC = NULL, precursorPpm = 10, ret = 10)
}
\arguments{
\item{mz}{mass to charge ratio (from the MS1 peak table).}

\item{RT}{retention time in seconds (from the MS1 peak table).}

\item{precursorPpm}{parts per million mass accuracy for match (default}

\item{ret}{retention time window (+/- seconds) for match}

\item{mz.v}{numeric vector of mass to charge ratio (from MS2 precursors)}

\item{RT.v}{numeric vector of retention time in seconds (from MS2 precursors)}
}
\value{
returns indices of matches as a list object
}
\description{
match ms1 features defined by mz and Rt to vector of ms2 features
}

