% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/publishApp.R
\name{publishApp}
\alias{publishApp}
\title{Publish your compMS2explorer app on shinyapps.io}
\usage{
publishApp(object, appName = NULL, ...)
}
\arguments{
\item{object}{a compMS2 class object}

\item{appName}{a name for your new app.}

\item{...}{further arguments to the \code{\link{shinyapps::deployApp}} function}
}
\description{
you must hold a shinyapps.io account and have a token.
use this command from the package rsconnect to set up your token
\code{rsconnect::setAccountInfo(name='your_shinyapps.io_user_name', token='your_token', secret='your_secret')}
Following this the app can be deployed on your account
}

