% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/runGitHubApp.R
\name{runGitHubApp}
\alias{runGitHubApp}
\title{run github shiny modified from shiny and devtools}
\usage{
runGitHubApp(repo = NULL, subdir = NULL, dirPath = NULL,
  auth_token = NULL, browserLaunch = TRUE)
}
\arguments{
\item{repo}{character github username and repository name. in the form "username/repositoryName"}

\item{subdir}{character sub-directory of the repo containing the shiny and data.}

\item{dirPath}{character full-path to a directory in which to save the contents of the zip file. If unsupplied shiny app will be opened from a temporary directory.}

\item{auth_token}{character private repo authorization token.}

\item{browserLaunch}{logical launch app in web browser (default = TRUE).}
}
\description{
run github shiny modified from shiny and devtools
}

