% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subsetCompMS2.R
\name{subsetCompMS2}
\alias{subsetCompMS2}
\title{subset compMS2 class object using a vector of spectra names}
\usage{
subsetCompMS2(object, ...)
}
\arguments{
\item{object}{a "compMS2" class object.}

\item{specNames}{character vector of composite spectrum names.}

\item{corrNetworkNodes}{logical should all the first correlation network 
nodes of the composite spectrum names also be returned? 
(default = FALSE).}
}
\value{
a "compMS2" class object with the composite spectra and all metID 
information removed.
}
\description{
subset compMS2 class object using a vector of spectra names
}
