% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subStructure.prob.R
\name{subStructure.prob}
\alias{subStructure.prob}
\title{Identifies probable substructure type}
\usage{
subStructure.prob(object, ...)
}
\arguments{
\item{object}{compMS2 object}

\item{minSumRelInt}{numeric (default = 70) miminum summed relative intensity to consider a probable
substructure type identification. If above this minimum summed relative intensity
then the most probable substructure type will be added to the compound_class 
column of the Comments table in the compMS2 object with a note stating it was
identified using this function e.g. sulfate (subStructure.prob). This provides
a means to automatically annotate the Comments table in a first-pass 
metabolite identification workflow.}
}
\value{
a data.frame of probable substructure annotations for each composite
spectrum, ranked by the sum of the relative intensities for that substructure
type
}
\description{
Identifies probable substructure type based on the summed
relative intensites and therefore proportion of total composite spectrum 
intensity explained.
}
